+++ lib/flop_phoenix/pagination.ex
@@ -30,15 +30,18 @@ defmodule Flop.Phoenix.Pagination do
     |> Keyword.put_new(:aria, label: "pagination")
   end
 
-  def build_page_link_helper(meta, route_helper, route_helper_args) do
+  def build_page_link_helper(meta, route_helper, route_helper_args, opts) do
     query_params =
-      meta.flop |> ensure_page_based_params() |> Flop.Phoenix.to_query()
+      meta.flop |> ensure_page_based_params() |> Flop.Phoenix.to_query(opts)
 
     fn page ->
-      params = Keyword.put(query_params, :page, page)
+      params = maybe_put_page(query_params, page)
       Flop.Phoenix.build_path(route_helper, route_helper_args, params)
     end
   end
+
+  defp maybe_put_page(params, 1), do: Keyword.delete(params, :page)
+  defp maybe_put_page(params, page), do: Keyword.put(params, :page, page)
 
   defp ensure_page_based_params(%Flop{} = flop) do
     # using default_limit without passing a page parameter produces a Flop